refactor: Orders 요소 폴더 분리
- type, components, mockData 대상 분리
