+++ src/features/orders/data/ordersMockData.ts
@@ -1,0 +1,67 @@
+export const MOCK_ORDERS = [
+    {
+        deliveryStatus: 'PREPARING_PRODUCT',
+        courierCompany: 'CJ 대한통운',
+        trackingNumber: '1234567890',
+        orderNumber: 'ORDER-20240921-00001',
+        paymentAt: '2025-11-08T11:54:18.375192088',
+        totalPaid: 200000,
+        recipientName: '홍길동',
+        orderItems: [
+            {
+                boardTitle: '예제 1',
+                itemName: '예제 2',
+                quantity: 2,
+                unitPrice: 50000,
+                totalPrice: 100000,
+                orderStatus: 'PAID',
+            },
+            {
+                boardTitle: '쿠키 1',
+                itemName: '쿠키 2',
+                quantity: 1,
+                unitPrice: 50000,
+                totalPrice: 100000,
+                orderStatus: 'PAID',
+            },
+        ],
+    },
+    {
+        deliveryStatus: 'SHIPPING',
+        courierCompany: '우체국 택배',
+        trackingNumber: '0987654321',
+        totalPaid: 100000,
+        orderNumber: 'ORDER-20240921-00002',
+        paymentAt: '2025-11-08T10:54:18.377200473',
+        recipientName: '김철수',
+        orderItems: [
+            {
+                boardTitle: '식빵',
+                itemName: '식빵',
+                quantity: 2,
+                unitPrice: 50000,
+                totalPrice: 100000,
+                orderStatus: 'PAID',
+            },
+        ],
+    },
+    {
+        deliveryStatus: 'PREPARING_PRODUCT',
+        courierCompany: null,
+        trackingNumber: null,
+        totalPaid: 150000,
+        orderNumber: 'ORDER-20240921-00003',
+        paymentAt: '2025-11-09T14:30:22.123456789',
+        recipientName: '이영희',
+        orderItems: [
+            {
+                boardTitle: '케이크',
+                itemName: '초코케이크',
+                quantity: 1,
+                unitPrice: 150000,
+                totalPrice: 150000,
+                orderStatus: 'PAID',
+            },
+        ],
+    },
+]
